Brings in upstream/main (through 8e196149127) and resolves the merge
conflict in DDLStmtExecutor.visitDropTaskStmt by combining:
- upstream's PIPE protection (#72141): rejects DROP TASK on PIPE-source
  tasks unless FORCE is set, alongside the existing MV-source check.
- this branch's TOCTOU re-check: mirrors the analyzer's creator-equality
  / OPERATE-on-SYSTEM rule on the resolved task right before
  taskManager.dropTasks(), so a concurrent SUBMIT TASK between analyze
  and execute cannot let an unprivileged caller drop someone else's
  task.

Both source-type checks now sit inside one `if (!isForce())` block, then
the auth re-check runs on the resolved task. PrivilegeCheckerTest still
passes (4 DROP-TASK cases including testDropTaskStmtExecutorReChecksAfterAnalyzerSkip).

+++ be/src/exec/pipeline/fetch_processor.cpp
@@ -227,8 +227,7 @@ StatusOr<FetchTaskPtr> FetchProcessor::_create_fetch_task(TupleId request_tuple_
         return std::make_shared<FetchTask>(std::move(task_ctx));
     }
     default:
-        return Status::InternalError(
-                fmt::format("Unknown row position type: {}", static_cast<uint8_t>(row_position_type)));
+        return Status::InternalError(fmt::format("Unknown row position type: {}", row_position_type));
     }
 }
